mqtt/rpi_requirements.txt
google_api_python_client==2.42.0
influxdb==5.3.1
paho_mqtt==1.6.1
protobuf==3.20.0
sense_hat==2.2.0
